Split up the pagination reducer code
