+++ src/app/store/reducers/pagination-reducer/pagination-reducer.types.ts
@@ -1,0 +1,5 @@
+
+
+export const resultPerPageParam = 'results-per-page';
+export const resultPerPageParamDefault = 5;
+
